requirements.txt
astroid==1.4.9
decorator==4.0.10
defusedxml==0.5.0rc1
Django==1.10.5
django-allauth==0.30.0
django-braces==1.11.0
django-cors-headers==2.1.0
django-facebook==6.0.3
django-filter==1.0.1
django-oauth-toolkit==0.11.0
django-rest-auth==0.9.0
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
geoindex==0.0.1
geopy==1.11.0
gunicorn==19.6.0
HTMLParser==0.0.2
imgurpython==1.1.7
isort==4.2.5
LatLon==1.0.2
lazy-object-proxy==1.2.2
lxml==3.7.2
Markdown==2.6.7
mccabe==0.6.1
mysqlclient==1.3.10
oauthlib==1.1.2
praw==4.0.0b17
prawcore==0.2.1
psycopg2==2.7.3
Pygments==2.2.0
PyJWT==1.4.2
pylint==1.6.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
pyproj==1.9.5.1
python-dateutil==2.6.0
python-geohash==0.8.5
python3-openid==3.0.10
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==1.0.1
social-auth-core==1.1.0
tabulate==0.7.7
Unidecode==0.4.20
update-checker==0.11
vincenty==0.1.4
whitenoise==3.2.1
wrapt==1.10.8
